content/html/content/test/test_bug583533.html

<!DOCTYPE HTML>
<html>
    <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=583533
-->
    <head>
        <title>Test for Bug 583514</title>
        <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
        <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    </head>
    <body>
        <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=583533">Mozilla Bug 583533</a>
        <p id="display"></p>
        <div id="content">
            <div id="e" accesskey="a">
  </div>
</div>
<pre id="test">
<script type="application/javascript">

    /** Test for Bug 583533 **/

    netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

    var sbs = Components.classes["@mozilla.org/intl/stringbundle;1"].
    getService(Components.interfaces.nsIStringBundleService);
    var bundle = sbs.createBundle("chrome://global-platform/locale/platformKeys.properties");
    
    var shiftText = bundle.GetStringFromName("VK_SHIFT");
    var altText = bundle.GetStringFromName("VK_ALT");
    var controlText = bundle.GetStringFromName("VK_CONTROL");
    var metaText = bundle.GetStringFromName("VK_META");
    var separatorText = bundle.GetStringFromName("MODIFIER_SEPARATOR");

    var modifier = Components.classes["@mozilla.org/preferences-service;1"]
                                      .getService(Components.interfaces.nsIPrefBranch)
                                      .getIntPref("ui.key.contentAccess");
                                      
    var domEvent = Components.interfaces.nsIDOMNSEvent;
    
    var isShift;
    var isAlt;
    var isControl;
    var isMeta;

    is(modifier < 16 && modifier >= 0, true, "Modifier in range");

    if (modifier & domEvent.META_MASK)
        isMeta = true;
    if (modifier & domEvent.SHIFT_MASK)
        isShift = true;
    if (modifier & domEvent.CONTROL_MASK)
        isControl = true;
    if (modifier & domEvent.ALT_MASK)
        isAlt = true;

    var label = "";

    if (isControl)
        label += controlText + separatorText;
    if (isMeta)
        label += metaText + separatorText;
    if (isAlt)
        label += altText + separatorText;
    if (isShift)
        label += shiftText + separatorText;

    label += document.getElementById("e").accessKey;

    is(label, document.getElementById("e").accessKeyLabel, "JS and C++ agree on accessKeyLabel");

</script>
</pre>
</body>
</html>
